PythonMasterAgent/requirements.txt
crewai
langchain
langchain-groq
langchain-together
langchain-huggingface
langchain-openai
python-telegram-bot
python-dotenv
pytz
schedule
tenacity